move delete methods to services
